requirements.txt
asgiref==3.7.2
attrs==23.1.0
better-profanity==0.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-dotenv==1.4.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
mysql==0.0.3
mysqlclient==2.2.0
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
profanity-check==1.0.3
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
scikit-learn==1.3.2
scipy==1.11.3
sqlparse==0.4.4
threadpoolctl==3.2.0
uritemplate==4.1.1
urllib3==2.0.7
